chore: Make pretty-printing an internal feature
