+++ xml-conduit/xml-conduit.cabal
@@ -36,6 +36,7 @@ library
                    , deepseq                   >= 1.1.0.0
     exposed-modules: Text.XML.Stream.Parse
                      Text.XML.Stream.Render
+                     Text.XML.Stream.Render.Internal
                      Text.XML.Unresolved
                      Text.XML.Cursor
                      Text.XML.Cursor.Generic
